fix: CLIN-984 merge frequency and clinical panel in summary panel
