--- src/views/screens/variant/Entity/ClinicalPanel/index.module.scss
@@ -1,6 +1,0 @@
-.clinicalPanelWrapper {
-  display: flex;
-  .clinicalPanel {
-    width: 100%;
-  }
-}
