--- src/views/screens/variant/Entity/FrequencyPanel/index.module.scss
@@ -1,7 +1,0 @@
-.frequencyPanelWrapper {
-    display: flex;
-    .frequencyPanel {
-      width: 100%;
-    }
-  }
-
