+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/conditions/CosmicConditionCell.tsx
