+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/conditions/DddConditionCell.tsx
