+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/conditions/HpoConditionCell.tsx
