+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/conditions/OmimConditionCell.tsx
