+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/conditions/OrphanetConditionCell.tsx
