+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/index.module.scss
@@ -1,0 +1,4 @@
+.clinicalCard {
+  width: 100%;
+  display: flex;
+}
