+++ src/views/screens/variant/Entity/SummaryPanel/Clinical/utils.tsx
