+++ src/views/screens/variant/Entity/SummaryPanel/Frequency/index.module.scss
@@ -1,0 +1,4 @@
+.frequencyCard {
+  display: flex;
+  width: 100%;
+}
